workers/profiling/requirements.txt
rq
psycopg2-binary
minio
